+++ app/src/main/java/eu/qm/fiszki/ActivityContainer/AddWordActivity.java
@@ -1,18 +1,23 @@
-package eu.qm.fiszki;
+package eu.qm.fiszki.ActivityContainer;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.EditText;
+
+import eu.qm.fiszki.DataBaseContainer.DBAdapter;
+import eu.qm.fiszki.DataBaseContainer.DBOperations;
+import eu.qm.fiszki.DataBaseContainer.DBStatus;
+import eu.qm.fiszki.R;
 
 public class AddWordActivity extends AppCompatActivity {
 
     EditText inputWord, inputTranslation;
     DBAdapter myDb = new DBAdapter(this);
-    OpenDataBaseClass OpenDataBase= new OpenDataBaseClass();
+    DBOperations myDbo;
+    DBStatus OpenDataBase = new DBStatus();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,21 +40,13 @@ public class AddWordActivity extends AppCompatActivity {
         int id = item.getItemId();
 
         if (id == R.id.action_add_new_word) {
-            if(!TextUtils.isEmpty(inputWord.getText()) || !TextUtils.isEmpty(inputTranslation.getText()))
-            {
-                myDb.insertRow(inputWord.getText().toString(), inputTranslation.getText().toString());
+            if (!TextUtils.isEmpty(inputWord.getText()) || !TextUtils.isEmpty(inputTranslation.getText())) {
+                myDbo.insertRow(inputWord.getText().toString(), inputTranslation.getText().toString());
             }
             inputWord.setText(null);
             inputTranslation.setText(null);
-            //populateListView();
             finish();
         }
         return super.onOptionsItemSelected(item);
-    }
-
-    private void closeDB()
-    {
-        myDb = new DBAdapter(this);
-        myDb.close();
     }
 }
