+++ app/src/main/java/eu/qm/fiszki/ActivityContainer/MainActivity.java
@@ -1,21 +1,27 @@
-package eu.qm.fiszki;
+package eu.qm.fiszki.ActivityContainer;
 
 import android.content.Intent;
 import android.database.Cursor;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.ListView;
 import android.widget.SimpleCursorAdapter;
 
+import eu.qm.fiszki.DataBaseContainer.DBAdapter;
+import eu.qm.fiszki.DataBaseContainer.DBModel;
+import eu.qm.fiszki.DataBaseContainer.DBOperations;
+import eu.qm.fiszki.DataBaseContainer.DBStatus;
+import eu.qm.fiszki.R;
+import eu.qm.fiszki.TimerClass;
+
 
 public class MainActivity extends AppCompatActivity {
-    TimerClass timer = new TimerClass();
 
+    TimerClass timer = new TimerClass();
+    DBOperations myDbo;
     DBAdapter myDb = new DBAdapter(this);
-    OpenDataBaseClass openDataBase = new OpenDataBaseClass();
+    DBStatus openDataBase = new DBStatus();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -37,13 +43,12 @@ public class MainActivity extends AppCompatActivity {
 
             Intent myIntent = new Intent(MainActivity.this, AddWordActivity.class);
             startActivity(myIntent);
-
     }
 
     private void populateListView()
     {
-        Cursor cursor = myDb.getAllRows();
-        String[] fromFieldNames = new String[] {DBAdapter.KEY_WORD, DBAdapter.KEY_TRANSLATION};
+        Cursor cursor = myDbo.getAllRows();
+        String[] fromFieldNames = new String[] {DBModel.KEY_WORD, DBModel.KEY_TRANSLATION};
         int[] toViewIDs = new int[] {R.id.word, R.id.translation};
         SimpleCursorAdapter myCursorAdapter;
         myCursorAdapter = new SimpleCursorAdapter(getBaseContext(),
@@ -51,6 +56,4 @@ public class MainActivity extends AppCompatActivity {
         ListView myList = (ListView) findViewById(R.id.listView);
         myList.setAdapter(myCursorAdapter);
     }
-
-
 }
