+++ app/src/main/java/eu/qm/fiszki/DataBaseContainer/DBAdapter.java
@@ -1,0 +1,56 @@
+package eu.qm.fiszki.DataBaseContainer;
+
+import android.content.Context;
+import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
+
+public class DBAdapter {
+
+    private final Context context;
+    private DatabaseHelper myDBHelper;
+    private SQLiteDatabase db;
+
+    public DBAdapter open()
+    {
+        db = myDBHelper.getWritableDatabase();
+        return this;
+    }
+
+    public void close()
+    {
+        myDBHelper.close();
+    }
+
+    public DBAdapter(Context ctx)
+    {
+        this.context = ctx;
+        myDBHelper = new DatabaseHelper(context);
+    }
+
+    private static class DatabaseHelper extends SQLiteOpenHelper
+    {
+        DatabaseHelper(Context context)
+        {
+            super(context, DBModel.DATABASE_NAME, null, DBModel.DATABASE_VERSION);
+        }
+
+        @Override
+        public void onCreate(SQLiteDatabase _db)
+        {
+            _db.execSQL(DBModel.DATABASE_CREATE_SQL);
+        }
+
+        @Override
+        public void onUpgrade(SQLiteDatabase _db, int oldVersion, int newVersion)
+        {
+            Log.w(DBModel.TAG, "Upgrading application's database from version " + oldVersion
+                    + " to " + newVersion + ", which will destroy all old data!");
+
+            _db.execSQL("DROP TABLE IF EXISTS " + DBModel.DATABASE_TABLE);
+
+            onCreate(_db);
+        }
+    }
+}
+
