+++ app/src/main/java/eu/qm/fiszki/DataBaseContainer/DBStatus.java
@@ -1,0 +1,16 @@
+package eu.qm.fiszki.DataBaseContainer;
+
+import eu.qm.fiszki.DataBaseContainer.DBAdapter;
+
+public class DBStatus {
+
+    public void openDB(DBAdapter dataAdapter)
+    {
+        dataAdapter.open();
+    }
+
+    private void closeDB(DBAdapter dataAdapter)
+    {
+        dataAdapter.close();
+    }
+}
