+++ app/src/main/java/eu/qm/fiszki/NotificationsClass.java
@@ -1,6 +1,5 @@
 package eu.qm.fiszki;
 
-import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
@@ -8,6 +7,9 @@ import android.content.Intent;
 import android.media.RingtoneManager;
 import android.net.Uri;
 import android.support.v7.app.NotificationCompat;
+
+import eu.qm.fiszki.ActivityContainer.CheckActivity;
+
 public class NotificationsClass {
     public void CreateNotification(Context window,String notificationMessage,
                                    String notificationTitle) {
