--- app/src/main/java/eu/qm/fiszki/OpenDataBaseClass.java
@@ -1,8 +1,0 @@
-package eu.qm.fiszki;
-
-public class OpenDataBaseClass {
-    public void openDB(DBAdapter dataAdapter)
-    {
-        dataAdapter.open();
-    }
-}
