--- app/src/main/java/eu/qm/fiszki/PopulateListViewClass.java
@@ -1,7 +1,0 @@
-package eu.qm.fiszki;
-
-/**
- * Created by Siusiacz on 2015-10-10.
- */
-public class PopulateListViewClass {
-}
