+++ app/src/main/res/layout-v15/activity_add_word.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin" tools:context="eu.qm.fiszki.AddWordActivity">
+    android:paddingBottom="@dimen/activity_vertical_margin" tools:context="eu.qm.fiszki.ActivityContainer.AddWordActivity">
 
     <TextView
         android:layout_width="wrap_content"
