+++ app/src/main/res/layout-v21/activity_check.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="eu.qm.fiszki.CheckActivity">
+    tools:context="eu.qm.fiszki.ActivityContainer.CheckActivity">
 
 
     <TextView
