+++ app/src/main/res/menu/menu_check.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="eu.qm.fiszki.CheckActivity">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="eu.qm.fiszki.ActivityContainer.CheckActivity">
     <item android:id="@+id/action_OK" android:title="@string/action_OK"
         android:orderInCategory="100" app:showAsAction="always" />
 </menu>
